js/tests/2024/day13.test.ts
import {describe, expect, it} from "vitest";
import {readPuzzleInput} from "../../src/shared/read.input";
import {part1, part2} from "../../src/2024/day13/day13";

const example = `
Button A: X+94, Y+34
Button B: X+22, Y+67
Prize: X=8400, Y=5400

Button A: X+26, Y+66
Button B: X+67, Y+21
Prize: X=12748, Y=12176

Button A: X+17, Y+86
Button B: X+84, Y+37
Prize: X=7870, Y=6450

Button A: X+69, Y+23
Button B: X+27, Y+71
Prize: X=18641, Y=10279
`;
const input = readPuzzleInput('2024/day13.txt');

describe('AoC 2024, Day 13, Part 1', () => {
    it('checks the example', () => {
        expect(part1(example)).toBe(480)
    })

    it('check the input', () => {
        expect(part1(input)).toBe(36250)
    })
});

describe('AoC 2024, Day 13, Part 2', () => {
    it('checks the example', () => {
        expect(part2(example)).toBe(875318608908)

    })

    it('checks the input', () => {
        expect(part2(input)).toBe(83232379451012)
    })
});